made blender html output simpler
